# fasta: Changed from busy-wait to waitFor()