In conversations with the Bens, particularly BHarsh, I took an IOU to
see how much switching from a busy-wait to a more principled waitFor()
hurts performance.  It's reasonably minimal, so I made the switch.  In
addition to using supplied features and avoiding the taboo of
busy-waiting, this also reduces the need for the helper functions which
I eliminated.

While here, I renamed nextTask to nextTid to improve the symmetry
between the waitFor() and the write().