@bradcray 2016-12-17
@chapel-lang/perf-team : Note that I made the change from busy-waiting to waitFor() as it didn't hurt performance much, and it came up a lot in conversations.  I plan to merge this later today but am leaving it open for a bit in case anyone happens to be online and wants to take a peek.