@bradcray 2016-12-17
Note: Tested 100 perf trials on shootout machine to look for problems and found none.  Tested under bencher script against 1.14 and everything seemed fine.