@ronawho 2016-12-18
yay!

You can make [numTasks](https://github.com/chapel-lang/chapel/blob/master/test/release/examples/benchmarks/shootout/fasta.chpl#L13) be `here.maxTaskPar` now, right?

Seems to improve timings slightly (timed with the release version): 
 - 2.2 ->1.9s on shootout box
 - 1.3 -> 1.2s on chap04
 - seems to stabilize chapcs numbers to 1.3s but I didn't do that many trials so no promises