@bradcray 2016-12-18
If anything, I think we'd now want to make 'numTasks' simply be '3' -- that's the degree of parallelism that the computation supports and some parameter sweeps suggest that using more or less doesn't help and often hurts.  But I kept the here.numTasks value in because I think if you only had 1 or 2 cores (as on u32), you'd really only want 1 or 2 tasks to avoid oversubscription.

I didn't do any timings with the 1.14 release other than on sealcs01 which is a pretty noisy machine, and only saw things stay the same or get incrementally worse.  nightly graphs seem to be saying that as well (though I can't recall the magic incantation to bring up the submitted shootout graphs as timed against 1.14).