@bradcray 2016-12-18
Oh wait, I'm realizing that I think I mis-parsed your comment.  You were saying changing to here.maxTaskPar had those effects?  I thought you were saying that my change had those effects.  Last I checked, increasing the number of tasks only hurt performance on the shootout box, but it's been awhile -- I'll double check.  It may be that '4' rather than here.maxTaskPar would be a sweet spot for chapcs in that it's an even number (would divide across numa evenly) and I can't think there's any advantage to adding more tasks (though I'll check that as well).