@bradcray 2016-12-18
Here's what I'm seeing, sweeping the number of tasks from 1 to here.maxTaskPar:

* on shootout box, slight improvement for numTasks = 4 over 3, as you note (nice!  wasn't true earlier, but timings have changed a lot since then; makes sense to have retimed).  Varying blocksize across powers of 2, 1024 seems to be the sweet spot.

* on chapcs07, best timings are 1.5sec - 1.6 sec for numTasks = 4, 2, and 6 in roughly that order.  1 task is slower, as you'd expect.  For odd numbers of tasks, things are noisiest.  12 tasks vary from 2 to 10 seconds; 24 tasks vary from 5 to 7 secs

* on chap04, things continue to be more stable and I get best numbers at numTasks = 5, then 6, then 4, 7, 8 are all about the same (and all of these differences are probably in the noise).

Based on these observations, my thinking is to change the numTasks to min(4, here.maxTaskPar) since using more than 4 tasks doesn't seem to result in much of a positive improvement.

I wanted to correct one thing in my previous thinking:  I was thinking that because there are three stages in the pipeline that using more than three tasks was unlikely to ever be useful.  However, what I forgot to take into account is that only 2 of the stages require a lock and that the third, which doesn't, is the longest running.  That means that having more tasks could take up some of the slack in that task.  But it isn't longer than the sum of the other two by much, which is why I think we don't see great improvements after 4 tasks.

Anyway, curious for your thoughts.  I'd like to make one more change to this tonight and submit tomorrow.