@bradcray 2016-12-18
Did a second run on chapcs07 because I didn't see anything like your 1.3s using 24 tasks and saw the same thing: best time at 4 tasks (1.1sec) and 24 tasks is in the 6.5-13.6 range (3 trials).  Can you reproduce that 1.3sec number?  Do you get better times at 4 tasks also?