@ronawho 2016-12-18
I was on chapcs00 (chapcs07 was still running nightly testing when I was trying it out.) Let me rerun now