@ronawho 2016-12-18
Still seeing 1.3 for chapcs

```
tail perfdat/chapcs07/fasta-release.dat
#
# master (numTasks=3)
12/18/16 	7.954
12/18/16 	10.21
12/18/16 	5.809

# numTasks=here.maxTaskPar
12/18/16 	1.305
12/18/16 	1.366
12/18/16 	1.314

```