@ronawho 2016-12-18
Oh, just did 10 trials, and i'm starting to see the fluctuation. Guess I just got really "lucky" for 6 runs before:

```
chapcs07 (master) $ tail fasta-release.dat
# Date	real
12/18/16 	9.456
12/18/16 	1.377
12/18/16 	2.005
12/18/16 	1.387
12/18/16 	2.594
12/18/16 	11.62
```