@bradcray 2016-12-18
Weird.  And that was on chapcs07?  I didn't get anything close to that.

Are you OK with the min(4, here.maxTaskPar) going onto master tonight?  (and likely into the site tomorrow barring surprises in the nightlies)?