@ronawho 2016-12-19
yeah, no problem with `min(4, here.maxTaskPar)`