@bradcray 2016-12-19
Oh, that's interesting... I was thinking I was on a local share, but actually I was only on a geographically local share, not system-local.  I'm glad to hear you still saw variance in that mode, though.